+++ modules/backend/src/app/api/project/webpage/webpage.controller.ts
@@ -1,6 +1,9 @@
 import { Body, Controller, Get, HttpCode, Param, Post } from '@nestjs/common'
 import { ApiTags } from '@nestjs/swagger'
+import { User } from '@prisma/client'
 
+import { AuthorizationService } from '../../../core/authorization/authorization.service'
+import { AuthUser } from '../../../core/guards/decorators/AuthUser'
 import { UseAuthGuard } from '../../../core/guards/decorators/UseAuthGuard'
 import { PrismaService } from '../../../core/prisma/prisma.service'
 import {
@@ -20,11 +23,13 @@ import {
 @ApiTags('/project/:projectFriendlyId/webpage')
 @Controller('/project/:projectFriendlyId/webpage')
 export class WebpageController {
-  constructor(private prisma: PrismaService) {}
+  constructor(private prisma: PrismaService, private authorization: AuthorizationService) {}
 
   @HttpCode(200)
   @Get('/list')
-  async list(@Param() param: WebpageListParam): Promise<WebpageListRes> {
+  async list(@AuthUser() user: User, @Param() param: WebpageListParam): Promise<WebpageListRes> {
+    await this.authorization.assertUserBelongsProject(user, param.projectFriendlyId)
+
     const project = await this.prisma.project.findUniqueOrThrow({
       where: { friendlyId: param.projectFriendlyId }
     })
@@ -45,7 +50,13 @@ export class WebpageController {
 
   @HttpCode(200)
   @Post('/add')
-  async add(@Param() param: WebpageAddParam, @Body() body: WebpageAddBody): Promise<WebpageAddRes> {
+  async add(
+    @AuthUser() user: User,
+    @Param() param: WebpageAddParam,
+    @Body() body: WebpageAddBody
+  ): Promise<WebpageAddRes> {
+    await this.authorization.assertUserIsProjectAdmin(user, param.projectFriendlyId)
+
     const project = await this.prisma.project.findUniqueOrThrow({
       where: { friendlyId: param.projectFriendlyId }
     })
@@ -69,9 +80,12 @@ export class WebpageController {
   @HttpCode(200)
   @Post('/:webpageId/edit')
   async edit(
+    @AuthUser() user: User,
     @Param() param: WebpageEditParam,
     @Body() body: WebpageEditBody
   ): Promise<WebpageEditRes> {
+    await this.authorization.assertUserIsProjectAdmin(user, param.projectFriendlyId)
+
     const webpage = await this.prisma.webpage.update({
       where: { id: param.webpageId },
       data: {
@@ -90,7 +104,12 @@ export class WebpageController {
 
   @HttpCode(200)
   @Post('/:webpageId/delete')
-  async delete(@Param() param: WebpageDeleteParam): Promise<WebpageDeleteRes> {
+  async delete(
+    @AuthUser() user: User,
+    @Param() param: WebpageDeleteParam
+  ): Promise<WebpageDeleteRes> {
+    await this.authorization.assertUserIsProjectAdmin(user, param.projectFriendlyId)
+
     await this.prisma.webpage.deleteMany({
       where: {
         id: param.webpageId,
